move GameLevel, Npc, and Player to GameEngine
