+++ assets/js/adventureGame/GameEngine/Player.js
@@ -1,4 +1,4 @@
-import Character from "./GameEngine/Character.js";
+import Character from "./Character.js";
 
 // Define non-mutable constants as defaults
 const SCALE_FACTOR = 25; // 1/nth of the height of the canvas
