+++ assets/js/adventureGame/GameLevelAirport.js
@@ -1,6 +1,6 @@
 import GameEnvBackground from './GameEngine/GameEnvBackground.js';
-import Npc from './Npc.js';
-import Player from './Player.js';
+import Npc from './GameEngine/Npc.js';
+import Player from './GameEngine/Player.js';
 import GameControl from './GameEngine/GameControl.js';
 import HelpPanel from './HelpPanel.js';
 import Game from './Game.js';
