+++ assets/js/adventureGame/GameLevelDesert.js
@@ -1,7 +1,7 @@
 // To build GameLevels, each contains GameObjects from below imports
 import GamEnvBackground from './GameEngine/GameEnvBackground.js';
-import Player from './Player.js';
-import Npc from './Npc.js';
+import Player from './GameEngine/Player.js';
+import Npc from './GameEngine/Npc.js';
 import Quiz from './Quiz.js';
 import GameControl from './GameEngine/GameControl.js';
 import GameLevelStarWars from './GameLevelStarWars.js';
